refactor(core): Typen aus Core in zuständige Projekte verschieben
- ChatMessage → ghGPT.Ai.Ollama (OpenAI-Protokolltyp, kein Domain-Konzept)
- ChatViews → ghGPT.Ai (internal, nur in ChatContextBuilder genutzt)
- Core/Ai-Typen in Einzeldateien aufgeteilt (ChatEvent, TokenEvent,
  ToolExecutedEvent, ToolCall, ToolCallResponse, ToolDefinition,
  ToolFunction, ChatRequest)
- ToolModels.cs entfernt
