+++ src/ghGPT.Ai/ChatViews.cs
@@ -1,0 +1,9 @@
+namespace ghGPT.Ai;
+
+internal static class ChatViews
+{
+    public const string Changes = "changes";
+    public const string History = "history";
+    public const string Branches = "branches";
+    public const string PullRequests = "pull-requests";
+}
