+++ src/ghGPT.Ai/IChatContextBuilder.cs
@@ -1,3 +1,4 @@
+using ghGPT.Ai.Ollama;
 using ghGPT.Core.Ai;
 
 namespace ghGPT.Ai;
