+++ src/ghGPT.Core/Ai/ChatRequest.cs
@@ -1,0 +1,9 @@
+namespace ghGPT.Core.Ai;
+
+public class ChatRequest
+{
+    public string Message { get; set; } = string.Empty;
+    public string? RepoId { get; set; }
+    public string? Branch { get; set; }
+    public string? ActiveView { get; set; }
+}
