+++ src/ghGPT.Core/Ai/TokenEvent.cs
@@ -1,0 +1,3 @@
+namespace ghGPT.Core.Ai;
+
+public record TokenEvent(string Token) : ChatEvent;
