+++ src/ghGPT.Core/Ai/ToolCallResponse.cs
@@ -1,0 +1,8 @@
+namespace ghGPT.Core.Ai;
+
+public class ToolCallResponse
+{
+    public bool HasToolCalls { get; set; }
+    public IReadOnlyList<ToolCall> ToolCalls { get; set; } = [];
+    public string? Content { get; set; }
+}
