+++ src/ghGPT.Core/Ai/ToolDefinition.cs
@@ -1,0 +1,7 @@
+namespace ghGPT.Core.Ai;
+
+public class ToolDefinition
+{
+    public string Type { get; set; } = "function";
+    public ToolFunction Function { get; set; } = null!;
+}
