+++ src/ghGPT.Core/Ai/ToolExecutedEvent.cs
@@ -1,0 +1,7 @@
+namespace ghGPT.Core.Ai;
+
+public record ToolExecutedEvent(
+    string ToolName,
+    string DisplayArgs,
+    bool Success,
+    string Message) : ChatEvent;
