+++ src/ghGPT.Core/Ai/ToolFunction.cs
@@ -1,0 +1,8 @@
+namespace ghGPT.Core.Ai;
+
+public class ToolFunction
+{
+    public string Name { get; set; } = string.Empty;
+    public string Description { get; set; } = string.Empty;
+    public object Parameters { get; set; } = new { };
+}
